% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmeans.R
\name{aic_kmeans2}
\alias{aic_kmeans2}
\title{Function to calculate the AIC (Akaike Information Criterion) for a matrix and cluster membership}
\usage{
aic_kmeans2(x, cl)
}
\arguments{
\item{x}{matrix with samples as rows and features as columns}

\item{cl}{vector of cluster memberships for the samples}
}
\description{
Function to calculate the AIC (Akaike Information Criterion) for a matrix and cluster membership
}
