% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmeans.R
\name{bic_kmeans}
\alias{bic_kmeans}
\title{Function to calculate the BIC (Bayesian Information Criterion) for a kmeans object}
\usage{
bic_kmeans(fit)
}
\arguments{
\item{fit}{kmeans object returned from kmeans()}
}
\description{
Function to calculate the BIC (Bayesian Information Criterion) for a kmeans object
}
