% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmeans.R
\name{bic_kmeans2}
\alias{bic_kmeans2}
\title{Function to calculate the BIC (Bayesian Information Criterion) for a matrix and cluster membership}
\usage{
bic_kmeans2(x, cl)
}
\arguments{
\item{x}{matrix with samples as rows and predictors as columns}

\item{cl}{vector of cluster memberships for the samples}
}
\description{
Function to calculate the BIC (Bayesian Information Criterion) for a matrix and cluster membership
}
