% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmeans.R
\name{.compute.wss}
\alias{.compute.wss}
\title{Helper function for bic_kmeans2}
\usage{
.compute.wss(x, f)
}
\arguments{
\item{x}{matrix with samples as rows and features as columns}

\item{f}{vector of cluster memberships for the samples}
}
\description{
Helper function for bic_kmeans2
}
