% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmeans.R
\name{multickmeans}
\alias{multickmeans}
\title{Consensus K-means clustering over mutiple K}
\usage{
multickmeans(x, ks, n_rep = 50, p_pred = 1, p_samp = 1,
  save_kms = TRUE, hclust_options = list(), calc_bic = TRUE, ...)
}
\arguments{
\item{x}{matrix with samples in rows and features in columns}

\item{ks}{vector of cluster numbers}

\item{n_rep}{number of individual k means runs}

\item{p_pred}{proportion of predictors used in every k means run}

\item{p_samp}{proportion of samples used in every k means run}

\item{save_kms}{logical, (or 'minimize') determining whether the k means object should be saved. This can be very memory demanding, depending on n_rep and x. If 'minimize', the kmeans objects are saved without column names, saving memory.}

\item{hclust_options}{list of option passed to hclust, which is used to generate consensus clusters}

\item{calc_bic}{logical, determining whether the BIC (Bayesian Information Criterion) should be calculated for the k means runs}

\item{...}{arguments passed to kmeans}
}
\value{
multickmeans object. ckms contains the list of ckmeans objects, ks is the vector of ks, bics is the vector of BICs.
}
\description{
Consensus k means clustering for multiple ks
}
\details{
Runs several independent k means clustering steps for several K, and combines information from the different runs to calculate consensus clusters using hierarchical clustering. The hierarchical clustering is based on the proportion of runs in which each pair of samples is placed in the same cluster, interpreted as distance.
}
\author{
Tankred Ott
}
