% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ckmeans.R
\name{plot.ckmeans}
\alias{plot.ckmeans}
\title{Plot ckmeans object as distance matrix}
\usage{
\method{plot}{ckmeans}(x, col = NULL, ord = TRUE, col_cl = NULL, plot_colorbar = TRUE, ...)
}
\arguments{
\item{x}{cKmeans object}

\item{col}{vector of colors (optional)}

\item{ord}{vectors of indices for ordering the matrix (optional).}

\item{col_cl}{vector of colors for the clusters (optional)}

\item{plot_colorbar}{logical determining whether a color bar should be plotted}

\item{...}{further arguments passed to the class specific plotDist functions}
}
\description{
Plots the consensus matrix as distance matrix
}
\details{
Consensus plot for cKmeans
}
