% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ckmeans.R
\name{plotDist.ckmeans}
\alias{plotDist.ckmeans}
\title{Plot ckmeans object as distance matrix}
\usage{
\method{plotDist}{ckmeans}(
  x,
  col = NULL,
  ord = NULL,
  returnOrd = FALSE,
  plotCC = TRUE,
  col_cl = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{cKmeans object}

\item{col}{vector of colors (optional)}

\item{ord}{vectors of indices for ordering the matrix (optional).}

\item{returnOrd}{logical, determining whether the indices used to order the plot should be returned (useful for plotting multiple distance plots with the same order)}

\item{plotCC}{logical, determining whether the consensus cluster should be plotted}

\item{col_cl}{vector of colors for the clusters (optional)}

\item{main}{title of the plot}

\item{...}{further arguments passed to the class specific plotDist functions}
}
\description{
Plots the consensus matrix as distance matrix
}
\details{
Distance (and consensus cluster) plot for cKmeans
}
