% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ckmeans.R
\name{plotDist.matrix}
\alias{plotDist.matrix}
\title{Plot distance matrix with clusters}
\usage{
\method{plotDist}{matrix}(x, cl = NULL, col = NULL, col_cl = NULL,
  ord = NULL, returnOrd = FALSE, main = "", ...)
}
\arguments{
\item{x}{n*n matrix}

\item{cl}{cluster membership}

\item{col}{vector of colors (optional)}

\item{col_cl}{vector of colors for the clusters (optional)}

\item{ord}{vectors of indices for ordering the matrix (optional).}

\item{returnOrd}{logical determining whether the indices used to order the plot should be returned (useful for plotting multiple distance plots with the same order)}

\item{main}{plot title}

\item{...}{further arguments passed to the class specific plotDist functions}
}
\description{
Plots a distance matrix with color coded clusters
}
\details{
Function to plot a distance matrix
}
\author{
Tankred Ott
}
