% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ckmeans.R
\name{plotDist}
\alias{plotDist}
\title{Plot distance matrix}
\usage{
plotDist(x, ...)
}
\arguments{
\item{x}{object to plot}

\item{...}{further arguments passed to the class specific plotDist functions}
}
\description{
Generic plotting function for distances (distance matrices)
}
\author{
Tankred Ott
}
