% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ckmeans.R
\name{plotDist.matrix}
\alias{plotDist.matrix}
\title{Plot distance matrix with clusters}
\usage{
\method{plotDist}{matrix}(
  x,
  cl = NULL,
  value_range = NULL,
  ord = TRUE,
  col = NULL,
  col_cl = NULL,
  plot_colorbar = FALSE,
  ...
)
}
\arguments{
\item{x}{n*n (distance) matrix}

\item{cl}{vector determining cluster membership}

\item{value_range}{vector with two elements c(d_min, d_max) determining the possible value range within the matrix. By default this will be the range of values in x.}

\item{ord}{vectors of indices for ordering of the matrix or boolean determining whether the matrix should be ordered}

\item{col}{vector of colors for the distance matrix}

\item{col_cl}{vector of colors or color ramp function for the clusters}

\item{plot_colorbar}{logical determining whether a color bar should be plotted}

\item{...}{further arguments}
}
\description{
Plots a distance matrix with color coded clusters
}
\details{
Function to plot a distance matrix
}
\author{
Tankred Ott
}
